Fixed deprecated Preference activity The UI is a bit different, but it's not noticeable.
